Remove precision limit on calculated resolution
The resolution attribute to the GeoData objects was rounded to 1 decimal
point.  This appears to have been done so the value would render nicely
within the logger, however storing the value with just one decimal point
of precision is problematic especially when having to rescale data
sources.

This commit removes the truncating of precision in that float.