Allow for dsm to be made up of an integer dtype
If the dsm is read in with an integer dtype, the *= operator with fail
when there is an attempt to rescale the dsm values based on the unit
scaling, which is often 1.0 (float dtype).

This commit also makes a slight tweak for pre-commit settings to install
numpy against the correct python version.