Have better checks and error messages
This removes some of the assert checks for checking of the transform,
and replaces them with more specific error checking that will give
more appropriate error messages.  Furthermore, the transform object is
preserved in the Affine object instead of converted to a 2-dimensional
numpy array.

If the ransac method does not have a model that converges, a more
specific exception is raised with a relevant message.

Lastly, some numpy methods were replaced with the standard library
equivalents where appropriate