Adjust codem resolution calculation
Instead of having codem set the resolution to identify matches to as
the max of fnd.native_resolution, aoi.native_resolution and
config["MIN_RESOLUTION"] (which defaults to 1.0), it will now set
the default config["MIN_RESOLUTION"] to float('nan') and set the
resolution to the max resolution of the foundation and AOI. If a
non-finite resolution is specified in the config, that will be used.
If the specified resolution is higher than both fnd.native_resolution
and aoi.native_resolution, a warning is emitted.