Apply colors to VCD pointcloud (#54)
* Apply colors to VCD pointcloud

This change adds color to the point cloud VCD output.  The color is
based on new vs fled for the ground pointcloud only.  This uses
the maplotlib blues and reds colormaps, although it would be trivial
to create different ones.

One colormap is applied to points classified as fled, the other colormap
is applied to the points classified as new.

* Add TODOs for doing colormaps

* Use the colorscale parameter assigned to the first product

* Remove colorscale products parameter, add colormap command line argument.