unindent WebSocketHandler definition
This addresses the issue where there is an isinstance check in __del__
for the log handler that checks if the log handler is of type
WebSocketHandler, and the WebSocketHandler is defined inside the else
block that follows the attempt to import websocket and pythonjsonlogger.
Since the WebSocketHandler class definition does not use anything from
the websocket-client library directly, we can move the definition of it
to be on the top level indent.