Handle the case with merged PC files
When using a PDAL pipeline as one of the arguments for codem with
filters.merge, the list comprehension to construct the reader metadata
is in the form of a list, and not a dictionary.  This commit replaces
the list comprehension with a generic for loop over the metadata, and
if the reader metadata value is in the form of a list, the metadata
list is extended to (instead of appended).